+++ app/routes/index.js
@@ -1,7 +1,13 @@
 import api from './api';
+import category from './category';
 import core from './core';
+import search from './search';
+import torrent from './torrent';
 
 export {
     api,
-    core
+    category,
+    core,
+    search,
+    torrent
 };
